fixed merge stuff

+++ lib/pages/settings/donor/pages/donor_settings_page.dart
@@ -33,8 +33,7 @@ class DonorSettingsPage extends GetView<DonorSettingsPageController> {
               PeopleHelpedWidget(
                 peopleHelped:
                     controller.donorStatistics.totalBeneficiariesDonatedTo,
-                amountDonated:
-                    controller.donorStatistics.totalAmountDonated.toDouble(),
+                amountDonated: controller.donorStatistics.totalAmountDonated,
                 countries: controller.donorStatistics.totalCountriesDonatedTo,
               ),
               const Gap(16),
